api/eval/resize/requirements.txt
pydantic >= 2.0.0 # required for inference.sh
inferencesh # required for inference.sh
Pillow
